feat : 콘서트 매진 랭킹 업데이트 기능 구현(이벤트사용)

+++ src/main/java/kr/hhplus/be/server/concert/domain/SoldoutRank.java
@@ -21,7 +21,7 @@ public class SoldoutRank {
     private LocalDateTime ticketOpenedAt;
     private LocalDateTime soldoutAt;
     private int seatTotal;
-    private Double score;
+    private long score;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
